requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-genai==0.3.0
yfinance==0.2.36
pillow==10.2.0
python-multipart==0.0.9
